refactor, test and bench multiexp
