+++ halo2_proofs/Cargo.toml
@@ -24,7 +24,7 @@ all-features = true
 rustdoc-args = ["--cfg", "docsrs", "--html-in-header", "katex-header.html"]
 
 [[bench]]
-name = "arithmetic"
+name = "multiexp"
 harness = false
 
 [[bench]]
