+++ halo2_proofs/src/multicore.rs
@@ -2,7 +2,7 @@
 //! `halo2`. It's currently just a (very!) thin wrapper around [`rayon`] but may
 //! be extended in the future to allow for various parallelism strategies.
 
-pub use rayon::{current_num_threads, scope, Scope};
+pub use rayon::{current_num_threads, prelude, scope, Scope};
 
 pub(crate) fn log_threads() -> u32 {
     let num_threads = current_num_threads();
